requirements.txt
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
